Worked on a kind of smarter job handling system and worked out what should happen if the job execution errors

+++ apps/CineFinnServer/src/database.ts
@@ -258,11 +258,11 @@ async function createTables() {
             null: false,
         },
         failed_at: {
-            type: 'int',
+            type: 'BIGINT',
             null: true,
         },
         finished_at: {
-            type: 'int',
+            type: 'BIGINT',
             null: true,
         },
         data: {
